LayoutTests/fast/dom/HTMLAnchorElement/anchor-file-blob-download-includes-backslash.html

<!DOCTYPE html><!-- webkit-test-runner [ useFlexibleViewport=true allowTestOnlyIPC=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width">
<script src="../../../resources/ui-helper.js"></script>
<script>
onload = async () => {
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.setShouldLogDownloadCallbacks(true);
        testRunner.waitUntilDownloadFinished();
    }
    await runTest();
}
</script>
</head>
<body>
<p>The suggested filename above should be "*.png" and the download should succeed.</p>
<a id="blob-url" download="*\">File backed blob URL</a>
<script>
function click(elmt)
{
    if (!window.eventSender) {
        alert('Click the link to run the test.');
        return;
    }
    UIHelper.activateAt(elmt.offsetLeft + 5, elmt.offsetTop + 5);
}

async function runTest()
{
    file = await internals.asyncCreateFile("../resources/abe.png");
    var link = document.getElementById("blob-url");
    link.href = window.URL.createObjectURL(file);
    click(link);
}
</script>
</body>
</html>
