Safari-609.1.11.1/LayoutTests/fast/dom/HTMLAnchorElement/anchor-file-blob-download-includes-backslash-expected.txt

Download started.
Downloading URL with suggested filename "*\.png"
Download completed.
The suggested filename above should be "*\.png" and the download should succeed.

File backed blob URL
